+++ app/src/main/java/com/example/fitbattleandroid/ui/common/Bottom.kt
@@ -8,8 +8,10 @@ import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.example.fitbattleandroid.ui.navigation.Screen
 import com.example.fitbattleandroid.ui.theme.primaryContainerDarkMediumContrast
 
 @Composable
@@ -32,9 +34,29 @@ fun NormalBottom(
 }
 
 @Composable
+fun TransparentBottom(
+    onClick:() -> Unit,
+    content: @Composable RowScope.() -> Unit
+){
+    Button(
+        onClick = onClick,
+        shape = RoundedCornerShape(20.dp),
+        colors =
+        ButtonDefaults.buttonColors(
+            containerColor = Color.Transparent,
+        ),
+    ) {
+        content()
+    }
+}
+
+@Composable
 @Preview
 fun NormalBottomPreview(){
     NormalBottom({/*Todo*/}){
+        Text("Hello")
+    }
+    TransparentBottom({/*Todo*/}) {
         Text("Hello")
     }
 }
