+++ app/src/main/java/com/example/fitbattleandroid/ui/screen/LoginScreen.kt
@@ -4,10 +4,12 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Button
@@ -28,7 +30,15 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
+import com.example.fitbattleandroid.ui.common.Background
+import com.example.fitbattleandroid.ui.common.Body
 import com.example.fitbattleandroid.ui.common.CommonOutlinedTextField
+import com.example.fitbattleandroid.ui.common.Header
+import com.example.fitbattleandroid.ui.common.MinText
+import com.example.fitbattleandroid.ui.common.NormalBottom
+import com.example.fitbattleandroid.ui.common.NormalText
+import com.example.fitbattleandroid.ui.common.TitleText
+import com.example.fitbattleandroid.ui.common.TransparentBottom
 import com.example.fitbattleandroid.ui.navigation.Screen
 import com.example.fitbattleandroid.ui.theme.onPrimaryDark
 import com.example.fitbattleandroid.ui.theme.primaryContainerDarkMediumContrast
@@ -39,100 +49,34 @@ fun LoginScreen(navController: NavController) {
     var email by remember { mutableStateOf("") }
     var password by remember { mutableStateOf("") }
 
-    Box(
-        contentAlignment = Alignment.Center,
-        modifier =
-            Modifier
-                .fillMaxSize()
-                .background(primaryContainerLight),
-    ) {
-        Column(
-            verticalArrangement = Arrangement.Top,
-            horizontalAlignment = Alignment.CenterHorizontally,
-            modifier =
-                Modifier
-                    .fillMaxSize()
-                    .imePadding(),
-        ) {
-            Box(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .background(primaryContainerDarkMediumContrast)
-                        .padding(16.dp),
-                contentAlignment = Alignment.Center,
-            ) {
-                Text(
-                    text = "Share Fit",
-                    style =
-                        MaterialTheme.typography.headlineMedium.copy(
-                            fontWeight = FontWeight.Bold,
-                            color = onPrimaryDark,
-                        ),
-                )
-            }
+    Background {
+        Header {
+            Body {
 
-            Column(
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally,
-                modifier =
-                    Modifier
-                        .fillMaxSize()
-                        .imePadding()
-                        .padding(16.dp),
-            ) {
-                Text(
-                    text = "ログイン",
-                    modifier = Modifier.padding(bottom = 60.dp),
-                    color = onPrimaryDark,
-                    style = MaterialTheme.typography.headlineMedium.copy(fontWeight = FontWeight.Bold),
-                )
+                TitleText("ログイン")
+
+                Spacer(modifier = Modifier.size(50.dp))
 
                 CommonOutlinedTextField(
                     value = email,
                     label = "メールアドレス",
                     onValueChange = { email = it },
                 )
+
                 CommonOutlinedTextField(
                     value = password,
                     label = "パスワード",
                     onValueChange = { password = it },
                 )
 
-                Button(
-                    onClick = { navController.navigate("main") },
-                    shape = RoundedCornerShape(20.dp),
-                    colors =
-                        ButtonDefaults.buttonColors(
-                            containerColor = primaryContainerDarkMediumContrast,
-                        ),
-                    modifier =
-                        Modifier
-                            .width(200.dp),
-                ) {
-                    Text(
-                        text = "ログイン",
-                        color = onPrimaryDark,
-                    )
+                NormalBottom({ navController.navigate("main") }) {
+                    NormalText("ログイン")
                 }
-                Button(
-                    onClick = {
-                        navController.navigate(Screen.Regi.route) {
-                            popUpTo(Screen.Login.route) { inclusive = true }
-                            launchSingleTop = true
-                        }
-                    },
-                    shape = RoundedCornerShape(20.dp),
-                    colors =
-                        ButtonDefaults.buttonColors(
-                            containerColor = Color.Transparent,
-                        ),
+
+                TransparentBottom({navController.navigate(Screen.Regi.route) { popUpTo(Screen.Login.route) { inclusive = true }
+                        launchSingleTop = true }}
                 ) {
-                    Text(
-                        text = "新規登録の方はこちら",
-                        fontSize = 10.sp,
-                        color = onPrimaryDark,
-                    )
+                    MinText("新規登録の方はこちら")
                 }
             }
         }
